requirements.txt
fastapi
aiohttp
pydantic
requests
sqlalchemy
psycopg2-binary
python-dotenv
edge-tts
werkzeug
flask_sqlalchemy
anthropic
aioredis
quart
redis